fix: deserialize JobOfferReceived.etaMinutes as double to match backend contract (BUG-060)
The backend broadcasts JobOfferReceived with etaMinutes as an unrounded double
(from HaversineCalculator). The simulator's JobOfferPayload record typed it as
int, so System.Text.Json threw on any fractional value during SignalR hub-argument
binding — the decision engine was never called and the offer was silently lost.

Fix: one-line type change (int → double). Backed by an ADR-0011 captured-payload
contract test that deserializes a real camelCase wire fixture with etaMinutes: 12.7
through JsonSerializerDefaults.Web (the same path JsonHubProtocol uses), asserting
all nine fields individually. Fails Red against the old int model (JsonException /
FormatException on $.etaMinutes), passes Green after the fix. A second test guards
the whole-number regression. 179 tests passing. Live-confirmed active during
BUG-055's gate: bind errors gone, fractional-ETA offers now flow and are accepted.
ADR-0011 sibling of BUG-036.

+++ src/ServiceDelivery.Simulator/Models/JobOfferPayload.cs
@@ -9,4 +9,4 @@ public sealed record JobOfferPayload(
     double Latitude,
     double Longitude,
     double DistanceMiles,
-    int EtaMinutes);
+    double EtaMinutes);
